Move constant buffer syntax to shared project
